refactor(events): use aura-events as single source for orchestration sse
The aura crate carried its own OrchestrationStreamEvent while
aura-events (the documented shared SSE-type crate) held a parallel
definition that had drifted: plan_created carried task_count where the
server emits tasks, and iteration_complete carried three required
evaluation-era fields the server never emits. Both drifted variants
were unmatchable against real payloads, so the typed consumer contract
was silently broken.

Reconcile aura-events to the actual wire format and delete the local
enum in favor of a re-export, the same pattern the base AuraStreamEvent
already uses. The TaskContext and CompletionOutcome composition structs
move into aura-events with the enum (task identity and outcome fields
stay defined once and flattened into their variants). RoutingMode is
unified too, gaining the DirectAnswer variant run manifests persist.

Wire-format snapshot tests prove the emitted JSON is unchanged across
the swap, and a tolerance test covers deserializing old
task_count-era payloads.

Ref: #221
